[WIP] implement calculating logic. multiple available entries missing

+++ src/main/kotlin/de/shopnchop/provideRecipes/RecipeIngredientEntity.kt
@@ -2,6 +2,6 @@ package de.shopnchop.provideRecipes
 
 data class RecipeIngredientEntity(
     val name: String,
-    val amount: Int,
+    val amount: Double,
     val unit: String
 )
